Added Pythonic suggestions
I didn't change the action of the page, but I made things more
manageable. Here are some outlines of the changes:

1. Moved the HTML and embed_codes to configuration files
This way, you can keep most of the code separate from silly
things like formatting.

2. Added constant variables for filenames
Technically, these would be best kept in a *different* config
file, but one hurdle at a time.

3. Utilized .gitignore with an output_files directory
Usually you don't want to commit generated files like the
HTML file we create here. The .gitignore file tells Git to...
well, ignore any file in the output_files directory since it is
not important to our actual code.

4. Utilized __name__ == '__main__'
This way, your code will actually do stuff only when explicitly
called through Python by 'python Animal_crossing.py'. If you
were to instead import your file, the code would not automatically
display a new web page.

5. Broke down the functions into smaller functions.
Generally, smaller functions are easier to manage and test than
bigger ones. (The podcast Coding Blocks says that around 4ish lines
of actual actions is usually appropriate).

6. Removed the Zip function for some math.
That was a good use of zip, but there was a better way. I'll let you work
out what I've done.

7. The 'with' command
Notice that I open all my files with the 'with' keyword. This is a
context manager for a file, so that the file is automatically opened,
exists within the code block, then as soon as the code block is finished
the file is automatically closed. Definitely look into it.